Add option to subscribe to all attestation subnets.
